Replaces newlines with space in testLevelDescription
If the `testLevelDescription` of a `ProcessTransaction` contains
newlines, carriage returns, or line breaks, the generated test suite is
not runnable due to an error like this:

it('should respond 200 for "Will send `Authenticated` if authentication is succesful,
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

because a normal string in JS can't be multi-line.

This diff runs a replace over the `testLevelDescription` that replaces
these newlines with spaces (" "). A space, as opposed to just stripping
the newlines, keeps things readable in the `it` blocks.